Fix pacman's query package install status
Make python packages installable from pip only when using virtualenv

This commit addresses issue #363. A virtualenv item is added to the
prefix specific config file if the prefix is created with a virtualenv.
While installing packages to the virtualenv, pybombs will use pip to
install python specific packages and normal operation while it is not a
python specific package

virtualenv set to False for prefix with no virtualenv