config_file: Make OrderedDict the one and only data type
Prior to this, the PBConfigFile class would bleed implementation details
by returning ruamel.ordereddict classes, which is an artifact of using
ruamel.yaml. Python's built-on OrderedDict class is more appropriate and
generic and should be used, and it doesn't give nasty surprises when
trying to manipulate return values from PBConfigFile.get() or other
methods.