git: Enable use of --reference-if-able
When going newer than git 2.11, --reference will fail with submodules
when the git reference is a bare repo. To avoid PyBOMBS crashing, the
usage of --reference-if-able is the more lenient and useful switch, but
it also requires git 2.11. The git fetcher now checks the git version,
and enables git caches with the appropriate switch.