build doxygen set to true by default
We were having an issue where people weren't seeing anything in the documentation tab of a block in GRC, and it turns out it's because this was disabled by default.  Yes I know you can enable doxygen with a flag, but so many people are using pybombs with the standard "pybombs install gnuradio" as their go-to method for new installs, and it makes GR seem less mature when the documentation tabs are empty.